Google: put refresh tokens into use
Use the OAuth2.0 "web_server" mechanism, along with the non-standard
"access_type=offline" parameter, to make Google return us a refresh token
(along with the usual access token) when authenticating.
Also, add the "approval_prompt=force" (again, non standard) parameter when the
plugin is (re-)authenticating the account. This parameter, however, must not be
used by client applications (or it will make the OAuth dialog appear all the
time, even when the tokens are still valid).