Do not reveal Google secret API key.
This was used to get a refresh token; however, since we are achieving the same
goal (long lived sessions) by means of the session cookies, this is no longer
necessary.
We might want to revert this patch if at some point we get confirmation that
Google doesn't mind us sharing the secret API key; but while in doubt, let's
keep it out of the source code.