Google: add ForceClientAuthViaRequestBody flag
The next version of signon-plugin-oauth2 will use HTTP Basic authorization for
authorizing the client, as recommended by the OAuth 2.0 specification. However,
Google does not support it, and therefore we need to add this special flag to
force the oauth2 plugin to authenticate the client by passing ClientId and
ClientSecret in the request body. 
We don't need to update the dependency version, because older versions of
signon-plugin-oauth2 will simply ignore this flag.