Do the worst thing ever, so Google lets us sign in (#37)
This change overrides the user agent string for Google logins to Chrome
41 on Windows 7.

This causes users to be notified of a login to their account from a
Windows device when they sign in using system-settings.

We don't know why this works. We got the idea from this StackOverflow
question:
https://stackoverflow.com/questions/40591090/403-error-thats-an-error-error-disallowed-useragent

This is not the right way to do this. If we can find any other way to
make Google login work, this should be reverted immediately.

In testing this issue, we tried many user agents, ranging from Chrome 40
to 99, Firefox, Internet Explorer, different operating systems, the
whole works... This singular Chrome 41 on Windows user agent is the only
one that works. I don't know why Google is blocking our sign-in
attempts, I don't know why this single user-agent is unblocked, and I
don't know how long this will work...

But it works for now.

Closes ubports/ubuntu-touch#1202